backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai==1.3.7
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
email_validator==2.1.0
chromadb==0.4.18
numpy<2.0.0
openai-whisper==20231117 